make cooldown default advanced function
